# apps/backend/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lead Cat API",
    "version": "1.0.0",
    "description": "Mini App-facing endpoints and health check for frontend codegen."
  },
  "paths": {
...
    "/api/auth/miniapp": {
      "post": {
        "operationId": "miniapp_auth_post",
        "tags": [
          "auth"
        ],
        "summary": "Exchange Telegram initData for Mini App JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MiniAppAuthRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MiniAppAuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Auth failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MiniAppAuthError"
                }
              }
            }
          }
        }
      }
    },
    "/api/miniapp/me": {
      "get": {
        "operationId": "miniapp_me_get",
        "tags": [
          "miniapp"
        ],
        "summary": "Current Mini App user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User identity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MiniAppUser"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/miniapp/meetings": {
      "get": {
        "operationId": "miniapp_meetings_get",
        "tags": [
          "miniapp"
        ],
        "summary": "List authed user meetings",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "scope",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "upcoming",
                "past",
                "all"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Meetings list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MiniAppMeetingsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "miniappCreateMeeting",
        "tags": [
          "miniapp"
        ],
        "summary": "Create a meeting (once or recurring series)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MiniAppMeetingCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created meeting",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "meeting"
                  ],
                  "properties": {
                    "meeting": {
                      "$ref": "#/components/schemas/MiniAppMeeting"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid body / validation_failed / meetings_not_configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "telegram_linked_to_other_account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/miniapp/schedule": {
      "get": {
        "operationId": "miniapp_schedule_get",
        "tags": [
          "miniapp"
        ],
        "summary": "Colleague schedule (read-only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "email"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "upcoming",
                "past",
                "all"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Meetings list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MiniAppMeetingsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/miniapp/employees": {
      "get": {
        "operationId": "miniapp_employees_get",
        "tags": [
          "miniapp"
        ],
        "summary": "Search employee directory",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Employees",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MiniAppEmployeesResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/miniapp/free-slots": {
      "post": {
        "operationId": "miniapp_free_slots_post",
        "tags": [
          "miniapp"
        ],
        "summary": "Common free-time checker",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MiniAppFreeSlotsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Free slots",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MiniAppFreeSlotsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/miniapp/meetings/{id}": {
      "patch": {
        "operationId": "miniappUpdateMeeting",
        "tags": [
          "miniapp"
        ],
        "summary": "Edit a single meeting (organizer/owner only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "this",
                "whole"
              ],
              "default": "this"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MiniAppMeetingUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated meeting",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "meeting"
                  ],
                  "properties": {
                    "meeting": {
                      "$ref": "#/components/schemas/MiniAppMeeting"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid body / invalid time / meetings_recurring_unsupported",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "miniappDeleteMeeting",
        "tags": [
          "miniapp"
        ],
        "summary": "Cancel a single meeting (organizer/owner only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "this",
                "whole"
              ],
              "default": "this"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Meeting cancelled"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/miniapp/admin/chat/link": {
      "post": {
        "operationId": "miniappAdminChatLink",
        "tags": [
          "miniapp-admin"
        ],
        "summary": "Link Telegram chat to workspace",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MiniAppAdminChatLinkRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Chat linked"
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/miniapp/admin/members/sync-chat": {
      "post": {
        "operationId": "miniappAdminMembersSyncChat",
        "tags": [
          "miniapp-admin"
        ],
        "summary": "Sync members from Telegram chat",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sync result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "added"
                  ],
                  "properties": {
                    "added": {
                      "type": "integer",
                      "format": "int64"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/miniapp/conflicts": {
      "post": {
        "operationId": "miniappConflicts",
        "tags": [
          "miniapp"
        ],
        "summary": "Cross-participant conflict warning",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MiniAppConflictsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Conflicts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "occurrences"
                  ],
                  "properties": {
                    "occurrences": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MiniAppOccurrenceConflicts"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid range / participants",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/miniapp/settings": {
      "get": {
        "operationId": "miniapp_settings_get",
        "tags": ["miniapp"],
        "summary": "Get current user settings",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Current settings",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MiniAppUserSettings" } } }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } }
          },
          "500": {
            "description": "Internal error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } }
          }
        }
      },
      "patch": {
        "operationId": "miniapp_settings_patch",
        "tags": ["miniapp"],
        "summary": "Update reminder minutes",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MiniAppUserSettingsPatch" } } }
        },
        "responses": {
          "204": { "description": "Saved" },
          "400": {
            "description": "validation_failed",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } }
          },
          "500": {
            "description": "Internal error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } }
          }
        }
      }
    },
    "/api/auth/web/{provider}/start": {
      "get": {
        "operationId": "webAuthProviderStart",
        "tags": ["web-auth"],
        "summary": "Begin OAuth SSO flow (redirect to provider)",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "enum": ["google", "microsoft"] }
          }
        ],
        "responses": {
          "302": { "description": "Redirect to OAuth provider" },
          "400": {
            "description": "Unknown provider",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } }
          }
        }
      }
    },
    "/api/auth/web/{provider}/callback": {
      "get": {
        "operationId": "webAuthProviderCallback",
        "tags": ["web-auth"],
        "summary": "OAuth SSO callback — exchange code for session cookie",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "enum": ["google", "microsoft"] }
          },
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "302": { "description": "Redirect to web dashboard (session cookie set)" },
          "400": {
            "description": "OAuth error / state mismatch",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } }
          }
        }
      }
    },
    "/api/auth/web/magic/request": {
      "post": {
        "operationId": "webAuthMagicRequest",
        "tags": ["web-auth"],
        "summary": "Request a magic-link email",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email"],
                "properties": {
                  "email": { "type": "string", "format": "email" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email sent (or silently accepted if address unknown)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "ok": { "type": "boolean" } }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } }
          }
        }
      }
    },
    "/api/auth/web/magic/verify": {
      "get": {
        "operationId": "webAuthMagicVerify",
        "tags": ["web-auth"],
        "summary": "Verify magic-link token and issue session cookie",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "302": { "description": "Redirect to web dashboard (session cookie set)" },
          "400": {
            "description": "Token missing",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } }
          },
          "401": {
            "description": "Token invalid or expired",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } }
          }
        }
      }
    },
    "/api/auth/web/logout": {
      "post": {
        "operationId": "webAuthLogout",
        "tags": ["web-auth"],
        "summary": "Invalidate current web session",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "204": { "description": "Session destroyed" },
          "401": {
            "description": "Not authenticated",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } }
          }
        }
      }
    },
    "/api/auth/web/me": {
      "get": {
        "operationId": "webAuthMe",
        "tags": ["web-auth"],
        "summary": "Get current web-session user identity",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "Authenticated user",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/WebUser" }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "web_session"
      }
    },
    "schemas": {
...
      "MiniAppAuthRequest": {
        "type": "object",
        "required": [
          "init_data"
        ],
        "properties": {
          "init_data": {
            "type": "string"
          }
        }
      },
      "MiniAppUser": {
        "type": "object",
        "required": [
          "telegram_id",
          "name",
          "email",
          "role"
        ],
        "properties": {
          "telegram_id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin"
            ]
          }
        }
      },
      "MiniAppAuthResponse": {
        "type": "object",
        "required": [
          "token",
          "user"
        ],
        "properties": {
          "token": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/MiniAppUser"
          }
        }
      },
      "MiniAppAuthError": {
        "type": "object",
        "required": [
          "code"
        ],
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "not_registered",
              "invalid_init_data"
            ]
          }
        }
      },
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        }
      },
...
    }
  }
}